refactor: add conversations and messages modules
