+++ src/views/app/detail-panel/conversations/conversation-preview-panel-container.tsx
@@ -12,19 +12,19 @@ import { filter, isEmpty } from 'lodash';
 import { useTranslation } from 'react-i18next';
 import { useNavigate, useParams } from 'react-router-dom';
 
+import { ConversationPreviewPanel } from './conversation-preview-panel';
+import { Spinner } from '../../../../assets/spinner';
+import { API_REQUEST_STATUS } from '../../../../constants';
+import { isFocusModeMailView } from '../../../../helpers/external-tabs';
+import { getFolderIdParts } from '../../../../helpers/folders';
+import { getConvEmailStoreAction } from '../../../../store/emails/actions/get-conv-action';
+import { useCompleteConversationOrFetch } from '../../../../store/emails/hooks/hooks';
+import { useConversationMessages } from '../../../../store/emails/store';
 import type {
 	DetailPanelConversationRouteParams,
 	DetailPanelRoutesParams
-} from '../../../types/routes';
-import { Spinner } from 'assets/spinner';
-import { API_REQUEST_STATUS } from 'constants/index';
-import { isFocusModeMailView } from 'helpers/external-tabs';
-import { getFolderIdParts } from 'helpers/folders';
-import { getConvEmailStoreAction } from 'store/emails/actions/get-conv-action';
-import { useCompleteConversationOrFetch } from 'store/emails/hooks/hooks';
-import { useConversationMessages } from 'store/emails/store';
-import { ConversationPreviewPanel } from 'views/app/detail-panel/conversation-preview-panel';
-import { PreviewPanelHeader } from 'views/app/detail-panel/preview/preview-panel-header';
+} from '../../../../types/routes';
+import { PreviewPanelHeader } from '../preview/preview-panel-header';
 
 export const ConversationPreviewPanelContainer = (): React.JSX.Element => {
 	const [t] = useTranslation();
