+++ src/views/app/detail-panel/messages/message-preview-panel.tsx
@@ -8,10 +8,10 @@ import React, { FC } from 'react';
 import { Container, Padding } from '@zextras/carbonio-design-system';
 import { useTranslation } from 'react-i18next';
 
-import { Spinner } from 'assets/spinner';
-import { IncompleteMessage, MailMessage } from 'types/index.d';
-import MailPreview from 'views/app/detail-panel/preview/mail-preview';
-import { PreviewPanelHeader } from 'views/app/detail-panel/preview/preview-panel-header';
+import { Spinner } from '../../../../assets/spinner';
+import { IncompleteMessage, MailMessage } from '../../../../types';
+import MailPreview from '../preview/mail-preview';
+import { PreviewPanelHeader } from '../preview/preview-panel-header';
 
 export const MessagePreviewPanel: FC<{
 	folderId: string;
