+++ src/views/app/detail-panel/tests/conversation-preview-panel-container.test.tsx
@@ -9,11 +9,11 @@ import React from 'react';
 import { act, screen } from '@testing-library/react';
 import * as shell from '@zextras/carbonio-shell-ui';
 
+import { ConversationPreviewPanelContainer } from '../conversations/conversation-preview-panel-container';
 import { setupTest } from '@test-setup';
 import { populateFoldersStore } from '@test-utils/store/folders';
 import { populateConversationInEmailStore } from '__test__/generators/generateConversation';
 import { updateConversationStatus } from 'store/emails/store';
-import { ConversationPreviewPanelContainer } from 'views/app/detail-panel/conversation-preview-panel-container';
 
 describe('ConversationPreviewPanelContainer', () => {
 	const defaultTitle = 'test title';
